refactor(types): layout 관련 type을 공용 types 디렉토리에서 types 하위로 이동
