+++ src/layout/ContentArea/ContentArea.tsx
@@ -11,7 +11,7 @@ import { v4 as uuid } from 'uuid'
 import useLayoutDispatch from '../hooks/useLayoutDispatch'
 import useMergeRefs from '../../hooks/useMergeRefs'
 import { CONTENT_MIN_WIDTH } from '../LayoutSizes'
-import ColumnType from '../../types/ColumnType'
+import ColumnType from '../types/ColumnType'
 import LayoutActions from '../redux/LayoutActions'
 import ContentAreaProps from './ContentArea.types'
 import { ContentAreaWrapper } from './ContentArea.styled'
