+++ src/layout/Side/SideArea/SideArea.styled.ts
@@ -1,6 +1,6 @@
 /* Internal dependencies */
 import { css, styled, disableAutoMinimum } from '../../../foundation'
-import LayoutSideType from '../../../types/LayoutSideType'
+import LayoutSideType from '../../types/LayoutSideType'
 
 interface SideAreaWrapperProps {
   sideType: LayoutSideType
