+++ src/layout/Side/SideArea/SideArea.types.ts
@@ -1,5 +1,5 @@
 /* Internal dependencies */
-import LayoutSideType from '../../../types/LayoutSideType'
+import LayoutSideType from '../../types/LayoutSideType'
 import { ChildrenComponentProps } from '../../../types/ComponentProps'
 
 export default interface SideAreaProps extends ChildrenComponentProps {
