+++ src/layout/Side/SideViewContent/SideViewContent.tsx
@@ -4,7 +4,7 @@ import { noop } from 'lodash-es'
 
 /* Internal dependencies */
 import useLayoutState from '../../hooks/useLayoutState'
-import LayoutSideType from '../../../types/LayoutSideType'
+import LayoutSideType from '../../types/LayoutSideType'
 import { SideArea } from '../SideArea'
 import SideViewContentProps from './SideViewContent.types'
