+++ src/layout/stories/LayoutPlayground/ContentHeader.tsx
@@ -5,7 +5,7 @@ import React from 'react'
 import { styled } from '../../../foundation'
 import { Text } from '../../../components/Text'
 import useHeader from '../../hooks/useHeader'
-import LayoutHeaderType from '../../../types/LayoutHeaderType'
+import LayoutHeaderType from '../../types/LayoutHeaderType'
 
 const Div = styled.div`
   display: flex;
